pyproject.toml
[tool.poetry]
name = "b81-spot"
version = "0.1.0"
description = "A script to check b81 spot availability and secure a spot if any"
authors = ["magnusja <github@mgns.tech>"]
license = "MIT" # chatgpts idea
repository = "https://github.com/magnusja/b81-spot"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
b81-spot = "check:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
